+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Cleaned up roxygen comments, DESCRIPTION, and NAMESPACE files.
 * Added a `NEWS.md` file to track changes to the package.
+* New `get_os()`, `is.os_mac`, `is.os_win`, `is.os_lnx`, and `is.os_unx` helper
+  functions to quickly determine operating system.